+++ src/dk/aau/cs/io/TapnXmlLoader.java
@@ -446,13 +446,13 @@ public class TapnXmlLoader {
 		boolean taggedArc = arc.getAttribute("tagged").equals("true") ? true : false;
 		String inscriptionTempStorage = arc.getAttribute("inscription");
 		String type = arc.getAttribute("type");
-		double nameOffsetXInput;
-		double nameOffsetYInput;
+		int nameOffsetXInput;
+		int nameOffsetYInput;
 		
 		//This check is done, as arcs in nets saved before this change do not have a nameOffset
 		if(arc.getAttribute("nameOffsetX") != "" && arc.getAttribute("nameOffsetY") != "") {
-			nameOffsetXInput = Double.parseDouble(arc.getAttribute("nameOffsetX"));
-			nameOffsetYInput = Double.parseDouble(arc.getAttribute("nameOffsetY"));
+			nameOffsetXInput = (int) Double.parseDouble(arc.getAttribute("nameOffsetX"));
+			nameOffsetYInput = (int) Double.parseDouble(arc.getAttribute("nameOffsetY"));
 		} else {
 			nameOffsetXInput = 0;
 			nameOffsetYInput = 0;
